fix and rename MongooseModelWrapper
