+++ app/api/odm/specs/mongooseModelWrapper_sessions.spec.ts
@@ -4,7 +4,7 @@ import { appContext } from 'api/utils/AppContext';
 import { testingEnvironment } from 'api/utils/testingEnvironment';
 import { ClientSession } from 'mongodb';
 import { Schema } from 'mongoose';
-import { MongooseModelWrapper } from '../MultiTenantMongooseModel';
+import { MongooseModelWrapper } from '../MongooseModelWrapper';
 
 interface TestDoc {
   title: string;
