+++ app/api/sync/syncsModel.ts
@@ -1,5 +1,5 @@
 import mongoose from 'mongoose';
-import { MongooseModelWrapper } from 'api/odm/MultiTenantMongooseModel';
+import { MongooseModelWrapper } from 'api/odm/MongooseModelWrapper';
 
 const syncSchema = new mongoose.Schema({
   lastSyncs: { type: mongoose.Schema.Types.Mixed, default: {} },
